xml/parsing/manakai/notations-1.dat

#data
<!DOCTYPE a[
<!N>
]><a/>
#errors
1;1;no XML decl;;;s
2;4;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!N
#errors
1;1;no XML decl;;;s
2;4;bogus comment;;;m
2;4;parser:EOF;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NO>
]><a/>
#errors
1;1;no XML decl;;;s
2;5;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NO
#errors
1;1;no XML decl;;;s
2;5;bogus comment;;;m
2;5;parser:EOF;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOT>
]><a/>
#errors
1;1;no XML decl;;;s
2;6;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOT
#errors
1;1;no XML decl;;;s
2;6;bogus comment;;;m
2;6;parser:EOF;;;m
2;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIO>
]><a/>
#errors
1;1;no XML decl;;;s
2;10;bogus comment;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOTATIO
#errors
1;1;no XML decl;;;s
2;10;bogus comment;;;m
2;10;parser:EOF;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION>
]><a/>
#errors
1;1;no XML decl;;;s
2;11;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION
#errors
1;1;no XML decl;;;s
2;11;parser:EOF;;;m
2;11;parser:EOF;;;m
2;11;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIONS>
]><a/>
#errors
1;1;no XML decl;;;s
2;11;no space between params;;;m
2;12;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION S "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATIONS
#errors
1;1;no XML decl;;;s
2;11;no space between params;;;m
2;12;parser:EOF;;;m
2;12;parser:EOF;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION b>
]><a/>
#errors
1;1;no XML decl;;;s
2;13;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b
#errors
1;1;no XML decl;;;s
2;13;parser:EOF;;;m
2;13;parser:EOF;;;m
2;13;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION b >
]><a/>
#errors
1;1;no XML decl;;;s
2;14;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "">
| <{} a>

#data escaped
<!DOCTYPE a[
<!NOTATION b\u0020
#errors
1;1;no XML decl;;;s
2;14;parser:EOF;;;m
2;14;parser:EOF;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd>
]><a/>
#errors
1;1;no XML decl;;;s
2;15;no md def;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc  #
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:string after name;;;m
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!NOTATION abc PUBLIC\u0020
#errors
1;1;no XML decl;;;s
2;23;parser:EOF;;;m
2;23;parser:EOF;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC #
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
2;24;parser:EOF;;;m
2;24;parser:EOF;;;m
2;24;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!NOTATION abc SYSTEM\u0020
#errors
1;1;no XML decl;;;s
2;23;parser:EOF;;;m
2;23;parser:EOF;;;m
2;23;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC #
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
2;24;parser:EOF;;;m
2;24;parser:EOF;;;m
2;24;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a"x
#errors
1;1;no XML decl;;;s
2;26;string after PUBLIC literal;;;m
2;27;parser:EOF;;;m
2;27;parser:EOF;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" x
#errors
1;1;no XML decl;;;s
2;27;string after PUBLIC literal;;;m
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM "a"x
#errors
1;1;no XML decl;;;s
2;26;string after SYSTEM literal;;;m
2;27;parser:EOF;;;m
2;27;parser:EOF;;;m
2;27;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc SYSTEM "a" x
#errors
1;1;no XML decl;;;s
2;27;string after SYSTEM literal;;;m
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" "b"x
#errors
1;1;no XML decl;;;s
2;30;string after SYSTEM literal;;;m
2;31;parser:EOF;;;m
2;31;parser:EOF;;;m
2;31;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION abc PUBLIC "a" "b" x
#errors
1;1;no XML decl;;;s
2;31;string after SYSTEM literal;;;m
2;32;parser:EOF;;;m
2;32;parser:EOF;;;m
2;32;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC>
]><a/>
#errors
1;1;no XML decl;;;s
2;22;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC >
]><a/>
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"
#errors
1;1;no XML decl;;;s
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg" "hij">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg""hij">
]><a/>
#errors
1;1;no XML decl;;;s
2;28;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg'"hij">
]><a/>
#errors
1;1;no XML decl;;;s
2;28;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"'hij'>
]><a/>
#errors
1;1;no XML decl;;;s
2;28;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg''hij'>
]><a/>
#errors
1;1;no XML decl;;;s
2;28;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "hij">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg" "hij"
#errors
1;1;no XML decl;;;s
2;34;parser:EOF;;;m
2;34;parser:EOF;;;m
2;34;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM>
]><a/>
#errors
1;1;no XML decl;;;s
2;22;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM >
]><a/>
#errors
1;1;no XML decl;;;s
2;23;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "efg">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"
#errors
1;1;no XML decl;;;s
2;28;parser:EOF;;;m
2;28;parser:EOF;;;m
2;28;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"x>
]><a/>
#errors
1;1;no XML decl;;;s
2;28;string after PUBLIC literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "efg" "">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "efg"x
#errors
1;1;no XML decl;;;s
2;28;string after PUBLIC literal;;;m
2;29;parser:EOF;;;m
2;29;parser:EOF;;;m
2;29;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC 'efg'x
#errors
1;1;no XML decl;;;s
2;28;string after PUBLIC literal;;;m
2;29;parser:EOF;;;m
2;29;parser:EOF;;;m
2;29;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"x>
]><a/>
#errors
1;1;no XML decl;;;s
2;28;string after SYSTEM literal;;;m
#document
| <!DOCTYPE a>
|   <!NOTATION bcd "" "efg">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM "efg"x
#errors
1;1;no XML decl;;;s
2;28;string after SYSTEM literal;;;m
2;29;parser:EOF;;;m
2;29;parser:EOF;;;m
2;29;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM 'efg'x
#errors
1;1;no XML decl;;;s
2;28;string after SYSTEM literal;;;m
2;29;parser:EOF;;;m
2;29;parser:EOF;;;m
2;29;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATION bcd SYSTEM 'efg'xyz<!
#errors
1;1;no XML decl;;;s
2;28;string after SYSTEM literal;;;m
2;33;parser:EOF;;;m
2;33;parser:EOF;;;m
2;33;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!NOTATIOn b SYSTEM "x">
]><a/>
#errors
1;1;no XML decl;;;s
2;3;keyword-wrong-case;;NOTATIOn;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!notation b SYSTEM "x">
]><a/>
#errors
1;1;no XML decl;;;s
2;3;keyword-wrong-case;;notation;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!nOTATION b SYSTEM "x">
]><a/>
#errors
1;1;no XML decl;;;s
2;3;keyword-wrong-case;;nOTATION;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b SYSTEM "x">
<!NOTATION b SYSTEM "y">
]><a/>
#errors
1;1;no XML decl;;;s
3;1;duplicate notation decl;;b;m
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION b SYSTEM "x">
%z;
<!NOTATION c SYSTEM "y">
]><a/>
#errors
1;1;no XML decl;;;s
3;1;entity not declared;;%z;;m
3;1;stop processing;;;i
#document
| <!DOCTYPE a>
|   <!NOTATION b "" "x">
|   <!NOTATION c "" "y">
| <{} a>

#data
<!DOCTYPE a[
<!NOTATION bcd PUBLIC "ab" 'efg'xy
#errors
1;1;no XML decl;;;s
2;33;string after SYSTEM literal;;;m
2;35;parser:EOF;;;m
2;35;parser:EOF;;;m
2;35;no root element;;;m
#document
| <!DOCTYPE a>
